fixes #3917 - replace protected_attrs with strong parameters
Filtering of attributes has moved from the protected_attributes gem to
strong parameters in controller concerns, to be in line with current
Rails recommendations.

Concerns are shared between UI and both API controllers and list the
attributes using Foreman::ParameterFilter, which provides additional
features:

1. Registration of additional attributes from plugins, through the
plugin API or `attr_accessible` for short term compatibility.
2. Re-use of permitted attribute lists for nested models (e.g. host
and interface).
3. Combining of lists of attributes from all sources in a single
permit call.
4. A small DSL for changing accepted parameters based on controller,
action and UI/API type.

Plugins should either temporarily depend on protected_attributes to
continue to protect their models or call permit/use ParameterFilter
similarly: http://projects.theforeman.org/projects/foreman/wiki/Strong_parameters

Some UI changes were required to make nested model hash keys all
integers instead of "new_123456" etc, else strong parameters would
filter the entries out.

The Role model's builtin default has been moved from initialize to the
database as the removal of protected_attrs changed the initialisation
order.

+++ app/controllers/concerns/foreman/controller/users_mixin.rb
@@ -2,8 +2,7 @@ module Foreman::Controller::UsersMixin
   extend ActiveSupport::Concern
 
   included do
-    before_action :set_admin_on_creation, :only => :create
-    before_action :clear_params_on_update, :update_admin_flag, :only => :update
+    before_action :clear_session_locale_on_update, :only => :update
   end
 
   def resource_scope(options = {})
@@ -12,38 +11,11 @@ module Foreman::Controller::UsersMixin
 
   protected
 
-  def set_admin_on_creation
-    admin = params[:user].delete :admin
-    @user = User.new(params[:user]) { |u| u.admin = admin unless admin.nil? }
-  end
-
-  def clear_params_on_update
-    if params[:user]
-      @admin = params[:user].has_key?(:admin) ? params[:user].delete(:admin) : nil
-      # Remove keys for restricted variables when the user is editing their own account
-      if editing_self?
-        params[:user].slice!(:password_confirmation,
-                             :password,
-                             :mail,
-                             :firstname,
-                             :lastname,
-                             :locale,
-                             :timezone,
-                             :default_organization_id,
-                             :default_location_id,
-                             :user_mail_notifications_attributes,
-                             :mail_enabled)
-
-        # Remove locale from the session when set to "Browser Locale" and editing self
-        session.delete(:locale) if params[:user][:locale].try(:empty?)
-      end
+  def clear_session_locale_on_update
+    if params[:user] && editing_self?
+      # Remove locale from the session when set to "Browser Locale" and editing self
+      session.delete(:locale) if params[:user][:locale].try(:empty?)
     end
-  end
-
-  def update_admin_flag
-    # Only an admin can update admin attribute of another user
-    # this is required, as the admin field is blacklisted above
-    @user.admin = @admin if User.current.admin && !@admin.nil?
   end
 
   def editing_self?
